Import heroku-buildpack-nginx.
This commit includes the heroku-buildpack-nginx as of September 8,
2023.

wwwhisper-buildpack initially used heroku-buildpack-nginx as a base.
However, after substantial changes, only the build_nginx script
remains in more or less its original form. Because of the large scope
of the changes, the wwwhisper-buildpack repository will not be
maintaned as a fork of heroku-buildpack-nginx and will not track its
history. Nevertheless, we include this initial commit to serve as a
reference for what has been changed in comparison to the
heroku-buildpack-nginx.